requirements.txt
Django==3.2.13
djangorestframework==3.13.1
pymongo==4.1.1